Moving sound functions to one SDLSound.c file, added fallback mode without sdl_mixer
