+++ externals/findsdl2-cmake/FindSDL2_mixer.cmake
@@ -98,3 +98,7 @@ set(SDLMIXER_INCLUDE_DIR ${SDL2_MIXER_INCLUDE_DIRS})
 set(SDLMIXER_FOUND ${SDL2_MIXER_FOUND})
 
 mark_as_advanced(SDL2_MIXER_LIBRARY SDL2_MIXER_INCLUDE_DIR)
+
+if(SDL2_MIXER_INCLUDE_DIR AND SDL2_MIXER_LIBRARY)
+    add_definitions(-DUSE_SDL_MIXER=1)
+endif()
