+++ src/MacSrc/Stub.c
@@ -1,37 +1,9 @@
 
 #include <Carbon/Carbon.h>
-#include "musicai.h"
-
-#include <SDL_mixer.h>
 
 // Stub out all of the functions that we haven't implemented yet!
 
-
-// Start with the sound library functions
-/*typedef struct {
-
-} snd_digi_parms;*/
-
 Boolean				gTuneDone;
-
-void snd_kill_all_samples(void) { }
-void snd_sample_reload_parms(snd_digi_parms *sdp) { }
-void snd_end_sample(int hnd_id) { }
-void snd_startup(void) { }
-
-int snd_start_digital(void) {
-	printf("snd_start_digital\n");
-	Mix_AllocateChannels(16);
-	return OK;
-}
-
-int snd_stop_digital(void) { return 1; }
-
-snd_digi_parms *snd_sample_parms(int hnd_id)
-{
-	snd_digi_parms parms;
-	return &parms;
-}
 
 void StopMovie(long x) { }
 void StartMovie(void * m) { }
